+++ OpenImis.RestApi/Controllers/PolicyControllerV1.cs
@@ -10,14 +10,15 @@ using OpenImis.RestApi.Security;
 
 namespace OpenImis.RestApi.Controllers
 {
+    [ApiVersion("1")]
     [Authorize]
     [Route("api/")]
     [ApiController]
     [EnableCors("AllowSpecificOrigin")]
-    public class PolicyController : Controller
+    public class PolicyControllerV1 : Controller
     {
         private readonly IImisModules _imisModules;
-        public PolicyController(IImisModules imisModules)
+        public PolicyControllerV1(IImisModules imisModules)
         {
             _imisModules = imisModules;
         }
