+++ OpenImis.RestApi/Docs/AddRequiredHeaderParameter.cs
@@ -1,0 +1,27 @@
+﻿using Swashbuckle.AspNetCore.Swagger;
+using Swashbuckle.AspNetCore.SwaggerGen;
+using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Threading.Tasks;
+
+namespace OpenImis.RestApi.Docs
+{
+    public class AddRequiredHeaderParameter : IOperationFilter
+    {
+        public void Apply(Operation operation, OperationFilterContext context)
+        {
+            if (operation.Parameters == null)
+                operation.Parameters = new List<IParameter>();
+
+            operation.Parameters.Add(new NonBodyParameter
+            {
+                Name = "api-version",
+                In = "header",
+                Type = "int",
+                Required = false,
+                Description = "API version"
+            });
+        }
+    }
+}
